requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
django-autoslug==1.9.8
django-autoslug-field==0.2.3
django-ckeditor==6.5.1
django-embed-video==1.4.8
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-social-share==2.3.0
idna==3.4
Pillow==9.4.0
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
